# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "CashMap API",
      "version": "1.0.0",
      "description": "API documentation for CashMap UMKM Cash Book Tracker"
    },
    "paths": {
      "/api/auth/register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                    "required": ["email", "password", "fullName", "securityAnswer", "businessName"],
                    "properties": {
                      "email": { "type": "string", "format": "email" },
                      "password": { "type": "string", "format": "password" },
                      "fullName": { "type": "string" },
                      "securityAnswer": { "type": "string" },
                      "businessName": { "type": "string" }
                    }
                }
              }
            }
          },
          "responses": {
            "201": { 
              "description": "User created successfully",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/User" }
                }
              }
            },
            "400": { "description": "Missing required fields" },
            "409": { "description": "Email already exists" }
          }
        }
      },
      "/api/auth/login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Login user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "password"],
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string", "format": "password" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/User" }
                }
              }
            },
            "401": { "description": "Invalid credentials" }
          }
        }
      },
      "/api/auth/profile": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get user profile",
          "responses": {
            "200": { 
              "description": "User profile data",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/User" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        },
        "put": {
          "tags": ["Auth"],
          "summary": "Update user profile",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": { "type": "string" },
                    "avatarUrl": { "type": "string" },
                    "businessName": { "type": "string" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { 
              "description": "Profile updated",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Auth"],
          "summary": "Reset Password",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "securityAnswer", "newPassword"],
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "securityAnswer": { "type": "string" },
                    "newPassword": { "type": "string", "format": "password" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Password reset successful" },
            "400": { "description": "Reset failed" }
          }
        }
      },
      "/api/transactions": {
        "get": {
          "tags": ["Transactions"],
          "summary": "Get all transactions",
          "responses": {
            "200": {
              "description": "List of transactions",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Transaction" }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        },
        "post": {
          "tags": ["Transactions"],
          "summary": "Create transaction",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["type", "amount", "category", "date"],
                  "properties": {
                    "type": { "type": "string", "enum": ["income", "expense"] },
                    "amount": { "type": "number" },
                    "category": { "type": "string" },
                    "description": { "type": "string" },
                    "date": { "type": "string", "format": "date" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { 
              "description": "Transaction created",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["Transactions"],
          "summary": "Update transaction",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": { "type": "string" },
                    "type": { "type": "string" },
                    "amount": { "type": "number" },
                    "category": { "type": "string" },
                    "description": { "type": "string" },
                    "date": { "type": "string", "format": "date" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Transaction updated" }
          }
        },
        "delete": {
          "tags": ["Transactions"],
          "summary": "Delete transaction",
          "parameters": [
            {
              "name": "id",
              "in": "query",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": { "description": "Transaction deleted" },
            "404": { "description": "Not found" }
          }
        }
      },
      "/api/backup": {
        "get": {
          "tags": ["Data"],
          "summary": "Export data",
          "responses": {
            "200": { "description": "JSON backup file" },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/import": {
        "post": {
          "tags": ["Data"],
          "summary": "Import data",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Backup JSON structure"
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Import successful" },
            "400": { "description": "Invalid file" }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": { "type": "string" },
            "email": { "type": "string", "format": "email" },
            "fullName": { "type": "string" },
            "businessName": { "type": "string", "nullable": true },
            "avatarUrl": { "type": "string", "nullable": true },
            "createdAt": { "type": "string", "format": "date-time" },
            "updatedAt": { "type": "string", "format": "date-time" }
          }
        },
        "Transaction": {
          "type": "object",
          "properties": {
            "id": { "type": "string" },
            "userId": { "type": "string" },
            "type": { "type": "string" },
            "amount": { "type": "number" },
            "category": { "type": "string" },
            "description": { "type": "string", "nullable": true },
            "date": { "type": "string", "format": "date-time" },
            "createdAt": { "type": "string", "format": "date-time" }
          }
        }
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }